Rename:  eo-h-f.mim -> eo-h-fundamente.mim, eo-h.mim -> eo-h-sistemo.mim, eo-q.mim -> eo-q-sistemo.mim, eo-x.mim -> eo-x-sistemo.mim
To make it easier to find the icons:

icons/eo-h-fundamente.png  icons/eo-h-sistemo.png  icons/eo-plena.png  icons/eo-q-sistemo.png  icons/eo-x-sistemo.png

When searching for the icon for eo-h-f.mim, a file eo-h-f.png is
searched and a file named according to the “(input-method eo
h-sistemo)” in the .mim file is searched, in this case
eo-h-sistemo.png. Therefore, ibus-m17n and ibus-typing-booster did
already find the icons before this rename and will still find the
icons after this rename. But in the documentation html generated by
m17n-docs, icons are only searched using the file name of the .mim
file. Therefore, in the website the icons for these input methods were
missing.