Use Shift+t (T) in ar-translit.mim to input the Arabic tatweel
I propose to add the Tatweel character to ar-translit.mim
The Tatweel character is used for justification in Arabic script.

On Windows and in ar-kbd.mim the character is produced using shift + j.
Shift + j in ar-translit.mim is already bound to "zero width joiner."
However, j is where ت is on Arabic keyboards which is the first letter
in Tatweel.

The patch therefore binds shift + t to Tatweel.

See also:

https://lists.nongnu.org/archive/html/m17n-list/2024-11/msg00000.html